Cleaned up code, refactored packages
