+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activities.HillfortMapsActivity"></activity>
+        <activity android:name=".views.map.HillfortMapView"></activity>
         <activity android:name=".views.hillfort.HillfortView" />
         <activity android:name=".views.hillfortlist.HillfortListView">
             <intent-filter>
@@ -27,7 +27,7 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name=".views.location.EditLocationView"
+            android:name=".views.editlocation.EditLocationView"
             android:label="@string/title_activity_map" />
     </application>
