+++ app/src/main/java/org/wit/placemark/views/editlocation/EditLocationView.kt
@@ -1,4 +1,4 @@
-package org.wit.placemark.views.location
+package org.wit.placemark.views.editlocation
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
