+++ app/src/main/java/org/wit/placemark/views/map/HillfortMapPresenter.kt
@@ -1,0 +1,34 @@
+package org.wit.placemark.views.map
+
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.model.LatLng
+import com.google.android.gms.maps.model.Marker
+import com.google.android.gms.maps.model.MarkerOptions
+import org.wit.placemark.main.MainApp
+
+class HillfortMapPresenter(val view: HillfortMapView) {
+
+    var app: MainApp
+
+    init {
+        app = view.application as MainApp
+    }
+
+    fun doPopulateMap(map: GoogleMap) {
+        map.uiSettings.setZoomControlsEnabled(true)
+        map.setOnMarkerClickListener(view)
+        app.hillforts.findAll().forEach {
+            val loc = LatLng(it.lat, it.lng)
+            val options = MarkerOptions().title(it.title).position(loc)
+            map.addMarker(options).tag = it.id
+            map.moveCamera(CameraUpdateFactory.newLatLngZoom(loc, it.zoom))
+        }
+    }
+
+    fun doMarkerSelected(marker: Marker) {
+        val tag = marker.tag as Long
+        val hillfort = app.hillforts.findById(tag)
+        if (hillfort != null) view.showHillfort(hillfort)
+    }
+}
