+++ app/src/main/java/org/wit/placemark/views/map/HillfortMapView.kt
@@ -1,0 +1,64 @@
+package org.wit.placemark.views.map
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.model.Marker
+import kotlinx.android.synthetic.main.activity_hillfort_maps.*
+import org.wit.placemark.R
+
+import org.wit.placemark.helpers.readImageFromPath
+import org.wit.placemark.models.HillfortModel
+
+class HillfortMapView : AppCompatActivity(), GoogleMap.OnMarkerClickListener {
+
+    lateinit var presenter: HillfortMapPresenter
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_hillfort_maps)
+        setSupportActionBar(toolbar)
+        presenter = HillfortMapPresenter(this)
+
+        mapView.onCreate(savedInstanceState);
+        mapView.getMapAsync {
+            presenter.doPopulateMap(it)
+        }
+    }
+
+    fun showHillfort(hillfort: HillfortModel) {
+        currentTitle.text = hillfort.title
+        currentDescription.text = hillfort.description
+        currentImage.setImageBitmap(readImageFromPath(this, hillfort.image))
+    }
+
+    override fun onMarkerClick(marker: Marker): Boolean {
+        presenter.doMarkerSelected(marker)
+        return true
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        mapView.onDestroy()
+    }
+
+    override fun onLowMemory() {
+        super.onLowMemory()
+        mapView.onLowMemory()
+    }
+
+    override fun onPause() {
+        super.onPause()
+        mapView.onPause()
+    }
+
+    override fun onResume() {
+        super.onResume()
+        mapView.onResume()
+    }
+
+    override fun onSaveInstanceState(outState: Bundle) {
+        super.onSaveInstanceState(outState)
+        mapView.onSaveInstanceState(outState)
+    }
+}
