+++ app/src/main/res/layout-land/activity_hillfort_maps.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.HillfortMapsActivity">
+    tools:context=".views.map.HillfortMapView">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
