+++ app/src/main/res/layout/activity_hillfort_maps.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.HillfortMapsActivity">
+    tools:context=".views.map.HillfortMapView">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
