+++ app/src/main/res/menu/menu_hillfort.xml
@@ -3,6 +3,15 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
+        android:id="@+id/item_save"
+        android:title="@string/menu_savePlacemark"
+        app:showAsAction="always"/>
+    <item
+        android:visible="false"
+        android:id="@+id/item_delete"
+        android:title="@string/menu_deleteHillfort"
+        app:showAsAction="always"/>
+    <item
         android:id="@+id/item_cancel"
         android:title="@string/menu_cancelHillfort"
         app:showAsAction="always"/>
