+++ app/src/main/res/values/strings.xml
@@ -21,4 +21,5 @@
     <string name="image2">Image2</string>
     <string name="image3">Image3</string>
     <string name="image4">Image4</string>
+    <string name="menu_savePlacemark">Save</string>
 </resources>
